+++ src/main/java/com/latebloomers/MovieInMine/config/SecurityConfig.java
@@ -1,0 +1,26 @@
+package com.latebloomers.MovieInMine.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@EnableWebSecurity // 스프링 시큐리티 필터가 스프링 필터체인에 등록된다.
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+        http.authorizeRequests()
+                .antMatchers("/", "/users").permitAll()
+                .anyRequest().authenticated()
+                .and()
+            .formLogin()
+                .loginPage("/users/signin")
+                .permitAll()
+                .and()
+            .logout()
+                .permitAll();
+    }
+}
