--- src/main/java/com/latebloomers/MovieInMine/model/Role.java
@@ -1,21 +1,0 @@
-package com.latebloomers.MovieInMine.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.List;
-
-@Entity
-@Getter @Setter
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
-    private String name;
-
-    @ManyToMany(mappedBy = "roles")
-    private List<User> users;
-}
