--- src/main/java/com/latebloomers/MovieInMine/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.latebloomers.MovieInMine.repository;
-
-import com.latebloomers.MovieInMine.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
-}
