--- src/main/java/com/latebloomers/MovieInMine/service/UserService.java
@@ -1,29 +1,0 @@
-package com.latebloomers.MovieInMine.service;
-
-import com.latebloomers.MovieInMine.model.Role;
-import com.latebloomers.MovieInMine.model.User;
-import com.latebloomers.MovieInMine.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-
-    public User save(User user){
-        String encodedPassword = passwordEncoder.encode(user.getPassword());
-        user.setPassword(encodedPassword);
-        user.setEnabled(true);
-        Role role = new Role();
-        role.setId(1);
-        user.getRoles().add(role);
-        return userRepository.save(user);
-    }
-
-}
