+++ src/main/resources/application.properties
@@ -11,3 +11,8 @@ mybatis.type-aliases-package=com.latebloomers.MovieInMine.model
 spring.mvc.pathmatch.matching-strategy = ANT_PATH_MATCHER
 
 server.port=8080
+
+spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jpa.show-sql=true
+
